Set Up Basic Twich Channel App
Set up the Client and Channel Components of the GetStream API and used
their documentation examples for setting up a functional running App
While running app initially with dummy channel members an error indicating
"role channel_member is not allowed to access Resource Create_Channel" is
encountered. To solve this simply go on the dashboard->lower nav menu->
click on chat->go to the channel type you set up while initializing the
channel for your app and click on it->Under the permissions section you
will see a huge array of action,resources,roles etc. select the item with
 role as channel_member and paste the string "Create_Channel" in the
 resources array at the end